refactor: Extract data handling
