+++ Cargo.toml
@@ -39,7 +39,7 @@ default = ["color-auto", "filesystem", "diff"]
 color = ["fs_snapshot/color"]
 color-auto = ["fs_snapshot/color-auto"]
 diff = ["fs_snapshot/diff"]
-filesystem = ["tempfile", "walkdir", "dunce", "content_inspector"]
+filesystem = ["tempfile", "walkdir", "dunce", "fs_snapshot/detect-encoding"]
 
 schema = ["schemars", "serde_json"]
 examples = ["escargot"]
@@ -71,7 +71,6 @@ wait-timeout = "0.2.0"
 tempfile = { version = "3.0", optional = true }
 walkdir = { version = "2.3.2", optional = true }
 dunce = { version = "1.0", optional = true }
-content_inspector = { version = "0.2.4", optional = true }
 
 schemars = { version = "0.8.3", features = ["preserve_order"], optional = true }
 serde_json = { version = "1.0", optional = true }
