+++ crates/fs_snapshot/Cargo.toml
@@ -38,6 +38,7 @@ harness = ["libtest-mimic", "ignore"]
 color = ["yansi", "concolor/std"]
 color-auto = ["color", "concolor/auto"]
 diff = ["difflib"]
+detect-encoding = ["content_inspector"]
 
 debug = ["backtrace"]
 
@@ -51,3 +52,4 @@ concolor = { version = "0.0.8", optional = true }
 
 backtrace = { version = "0.3", optional = true }
 difflib = { version = "0.4", optional = true }
+content_inspector = { version = "0.2.4", optional = true }
