+++ crates/fs_snapshot/src/lib.rs
@@ -1,4 +1,5 @@
 mod action;
+mod data;
 mod error;
 
 pub mod report;
@@ -8,5 +9,7 @@ pub mod utils;
 pub mod harness;
 
 pub use action::Action;
+pub use data::Data;
 pub use error::Error;
+
 pub type Result<T, E = Error> = std::result::Result<T, E>;
