+++ src/lib.rs
@@ -189,7 +189,9 @@ pub use cases::TestCases;
 pub use fs_snapshot::Error;
 
 pub(crate) use command::wait_with_input_output;
-pub(crate) use filesystem::{shallow_copy, Data, FilesystemContext, Iterate as FsIterate};
+pub(crate) use filesystem::{shallow_copy, FilesystemContext, Iterate as FsIterate};
 pub(crate) use registry::BinRegistry;
 pub(crate) use runner::{Case, Mode, Runner};
 pub(crate) use spec::RunnerSpec;
+
+pub(crate) use fs_snapshot::Data;
